refactor(macros): split SwiftUIMacros into per-macro files (N-1)
Separate the two SwiftUI-adjacent macros and the utilities they share so
each file has a single focus. No behavior change — `swift test` (349
tests) stays green and macro expansion snapshots are byte-identical.

- `DIEnvironmentBridgeMacro.swift` — MemberMacro + ExtensionMacro
  conformances plus `validateEnvironmentBridge` /
  `makeEnvironmentBridgeModifierDecl` / `makeEnvironmentBridgeHelperDecl`
  / `makeEnvironmentBridgeBodyExpr` and the two private validation model
  types.
- `DIFeatureRootMacro.swift` — PeerMacro conformance plus
  `parseFeatureRootAttribute`, `featureRootHelperName`, and
  `featureRootHelperConflicts` (the alias / duplicate / helper-name
  conflict resolution path).
- `SwiftUIMacros.swift` (kept, 722 → 216 lines) — shared AST helpers:
  `containerMemberNames`, `nominalTypeSyntax`, access-level mapping,
  `hasAttribute` / `featureRootAttributes`, feature-root alias
  validation, `enclosingDeclGroup`, and a couple of small utilities.
  Helpers promoted from `private` to `internal` so both macro files can
  call them.

Closes N-1 (the large-file split PR). Next PRs in the Phase N stack
start adding user-facing changes (N-2 Sendable hardening, N-3 child
overrides diagnostic, N-4 typealias warning, N-5 docs + bindings tests).
