+++ Sources/InnoDIMacros/SwiftUIMacros.swift
@@ -1,536 +1,22 @@
+//
+//  SwiftUIMacros.swift
+//  InnoDIMacros
+//
+//  Phase N-1 — thin helpers shared by `DIEnvironmentBridgeMacro.swift` and
+//  `DIFeatureRootMacro.swift`. The macro implementations themselves live in
+//  their own files now; this module keeps only the common AST utilities
+//  (container member enumeration, nominal-type synthesis, access-level
+//  modifier mapping, attribute lookup, feature-root alias validation, etc.).
+//
+
 import Foundation
 import InnoDICore
-import SwiftDiagnostics
 import SwiftSyntax
 import SwiftSyntaxBuilder
-import SwiftSyntaxMacros
 
-public struct DIEnvironmentBridgeMacro {}
+// MARK: - Container shape introspection
 
-extension DIEnvironmentBridgeMacro: MemberMacro {
-    public static func expansion(
-        of node: AttributeSyntax,
-        providingMembersOf declaration: some DeclGroupSyntax,
-        conformingTo protocols: [TypeSyntax],
-        in context: some MacroExpansionContext
-    ) throws -> [DeclSyntax] {
-        try expansion(of: node, providingMembersOf: declaration, in: context)
-    }
-
-    public static func expansion(
-        of node: AttributeSyntax,
-        providingMembersOf declaration: some DeclGroupSyntax,
-        in context: some MacroExpansionContext
-    ) throws -> [DeclSyntax] {
-        guard let nominalType = nominalTypeSyntax(for: declaration) else {
-            return []
-        }
-
-        let validation = validateEnvironmentBridge(
-            attribute: node,
-            declaration: declaration,
-            context: context,
-            emitDiagnostics: true
-        )
-        guard let mappings = validation.mappings else {
-            return []
-        }
-
-        let accessLevel = accessLevelModifiers(for: declaration.modifiers)
-        let modifierDecl = makeEnvironmentBridgeModifierDecl(
-            accessLevel: accessLevel,
-            containerType: nominalType,
-            mappings: mappings
-        )
-        let bridgeMethodDecl = makeEnvironmentBridgeHelperDecl(accessLevel: accessLevel)
-
-        return [DeclSyntax(modifierDecl), DeclSyntax(bridgeMethodDecl)]
-    }
-}
-
-extension DIEnvironmentBridgeMacro: ExtensionMacro {
-    public static func expansion(
-        of node: AttributeSyntax,
-        attachedTo declaration: some DeclGroupSyntax,
-        providingExtensionsOf type: some TypeSyntaxProtocol,
-        conformingTo protocols: [TypeSyntax],
-        in context: some MacroExpansionContext
-    ) throws -> [ExtensionDeclSyntax] {
-        let validation = validateEnvironmentBridge(
-            attribute: node,
-            declaration: declaration,
-            context: context,
-            emitDiagnostics: false
-        )
-        guard validation.mappings != nil else {
-            return []
-        }
-
-        return [
-            try ExtensionDeclSyntax("extension \(type): DIEnvironmentBridging {}")
-        ]
-    }
-}
-
-public struct DIFeatureRootMacro {}
-
-extension DIFeatureRootMacro: PeerMacro {
-    public static func expansion(
-        of node: AttributeSyntax,
-        providingPeersOf declaration: some DeclSyntaxProtocol,
-        in context: some MacroExpansionContext
-    ) throws -> [DeclSyntax] {
-        guard let varDecl = declaration.as(VariableDeclSyntax.self),
-              let binding = varDecl.bindings.first,
-              let identifier = binding.pattern.as(IdentifierPatternSyntax.self),
-              let typeAnnotation = binding.typeAnnotation,
-              let info = parseFeatureRootAttribute(node) else {
-            return []
-        }
-
-        guard hasAttribute(named: "SubContainer", in: varDecl.attributes) else {
-            context.diagnose(
-                Diagnostic(
-                    node: Syntax(node),
-                    message: SimpleDiagnostic.swiftUIFeatureRootWithoutSubContainer()
-                )
-            )
-            return []
-        }
-
-        let propertyName = identifier.identifier.text
-        if let alias = info.alias, !isValidFeatureRootAlias(alias) {
-            context.diagnose(
-                Diagnostic(
-                    node: Syntax(node),
-                    message: SimpleDiagnostic.swiftUIFeatureRootInvalidAlias(alias: alias)
-                )
-            )
-            return []
-        }
-        let helperName = featureRootHelperName(propertyName: propertyName, alias: info.alias)
-
-        let featureRootAttributes = featureRootAttributes(in: varDecl.attributes)
-
-        if info.alias == nil {
-            let aliaslessAttributes = featureRootAttributes.filter {
-                parseFeatureRootAttribute($0)?.alias == nil
-            }
-            if let currentIndex = aliaslessAttributes.firstIndex(where: { $0 == node }),
-               currentIndex > 0 {
-                context.diagnose(
-                    Diagnostic(
-                        node: Syntax(node),
-                        message: SimpleDiagnostic.swiftUIFeatureRootDuplicateDefault(propertyName: propertyName)
-                    )
-                )
-                return []
-            }
-        }
-
-        if let enclosingDecl = enclosingDeclGroup(containing: Syntax(declaration)),
-           featureRootHelperConflicts(
-                helperName: helperName,
-                currentAttribute: node,
-                in: enclosingDecl
-           ) {
-            context.diagnose(
-                Diagnostic(
-                    node: Syntax(node),
-                    message: SimpleDiagnostic.swiftUIFeatureRootHelperNameConflict(helperName: helperName)
-                )
-            )
-            return []
-        }
-
-        let accessLevel = accessLevelModifierText(for: enclosingDeclModifiers(containing: Syntax(declaration)))
-        let childTypeName = typeAnnotation.type.trimmedDescription
-        let rootViewTypeName = info.rootViewTypeName
-
-        let helperDecl: DeclSyntax = """
-            \(raw: accessLevel)func \(raw: helperName)() -> \(raw: rootViewTypeName) {
-                \(raw: rootViewTypeName)(container: \(raw: propertyName))
-            }
-            """
-
-        _ = childTypeName
-        return [helperDecl]
-    }
-}
-
-private struct EnvironmentBridgeMappingInfo {
-    let memberName: String
-    let environmentKeyPath: ExprSyntax
-}
-
-private struct EnvironmentBridgeValidationResult {
-    let mappings: [EnvironmentBridgeMappingInfo]?
-}
-
-private struct FeatureRootAttributeInfo {
-    let rootViewTypeName: String
-    let alias: String?
-}
-
-private func validateEnvironmentBridge(
-    attribute: AttributeSyntax,
-    declaration: some DeclGroupSyntax,
-    context: some MacroExpansionContext,
-    emitDiagnostics: Bool
-) -> EnvironmentBridgeValidationResult {
-    guard let arguments = attribute.arguments?.as(LabeledExprListSyntax.self),
-          let firstArgument = arguments.first,
-          let arrayExpr = firstArgument.expression.as(ArrayExprSyntax.self) else {
-        if emitDiagnostics {
-            context.diagnose(
-                Diagnostic(
-                    node: Syntax(attribute),
-                    message: SimpleDiagnostic.swiftUIEnvironmentBridgeInvalidArguments()
-                )
-            )
-        }
-        return EnvironmentBridgeValidationResult(mappings: nil)
-    }
-
-    let memberNames = Set(containerMemberNames(in: declaration))
-    var seenMembers: Set<String> = []
-    var mappings: [EnvironmentBridgeMappingInfo] = []
-    var hadErrors = false
-
-    for element in arrayExpr.elements {
-        guard let tupleExpr = element.expression.as(TupleExprSyntax.self) else {
-            if emitDiagnostics {
-                context.diagnose(
-                    Diagnostic(
-                        node: Syntax(element.expression),
-                        message: SimpleDiagnostic.swiftUIEnvironmentBridgeInvalidKeyPath(label: "member")
-                    )
-                )
-            }
-            hadErrors = true
-            continue
-        }
-
-        var memberName: String?
-        var environmentKeyPath: ExprSyntax?
-
-        for tupleElement in tupleExpr.elements {
-            guard let label = tupleElement.label?.text else {
-                continue
-            }
-
-            switch label {
-            case "member":
-                guard let parsedMemberName = stringLiteralValue(tupleElement.expression) else {
-                    if emitDiagnostics {
-                        context.diagnose(
-                            Diagnostic(
-                                node: Syntax(tupleElement.expression),
-                                message: SimpleDiagnostic.swiftUIEnvironmentBridgeInvalidKeyPath(label: "member")
-                            )
-                        )
-                    }
-                    hadErrors = true
-                    continue
-                }
-                memberName = parsedMemberName
-            case "environment":
-                guard let keyPath = tupleElement.expression.as(KeyPathExprSyntax.self) else {
-                    if emitDiagnostics {
-                        context.diagnose(
-                            Diagnostic(
-                                node: Syntax(tupleElement.expression),
-                                message: SimpleDiagnostic.swiftUIEnvironmentBridgeInvalidKeyPath(label: "environment")
-                            )
-                        )
-                    }
-                    hadErrors = true
-                    continue
-                }
-                environmentKeyPath = ExprSyntax(keyPath)
-            default:
-                continue
-            }
-        }
-
-        guard let memberName, let environmentKeyPath else {
-            hadErrors = true
-            continue
-        }
-
-        guard memberNames.contains(memberName) else {
-            if emitDiagnostics {
-                context.diagnose(
-                    Diagnostic(
-                        node: Syntax(element.expression),
-                        message: SimpleDiagnostic.swiftUIEnvironmentBridgeUnknownMember(memberName: memberName)
-                    )
-                )
-            }
-            hadErrors = true
-            continue
-        }
-
-        guard seenMembers.insert(memberName).inserted else {
-            if emitDiagnostics {
-                context.diagnose(
-                    Diagnostic(
-                        node: Syntax(element.expression),
-                        message: SimpleDiagnostic.swiftUIEnvironmentBridgeDuplicateMember(memberName: memberName)
-                    )
-                )
-            }
-            hadErrors = true
-            continue
-        }
-
-        mappings.append(
-            EnvironmentBridgeMappingInfo(
-                memberName: memberName,
-                environmentKeyPath: environmentKeyPath
-            )
-        )
-    }
-
-    return EnvironmentBridgeValidationResult(mappings: hadErrors ? nil : mappings)
-}
-
-private func makeEnvironmentBridgeModifierDecl(
-    accessLevel: DeclModifierListSyntax,
-    containerType: TypeSyntax,
-    mappings: [EnvironmentBridgeMappingInfo]
-) -> StructDeclSyntax {
-    let containerDecl = VariableDeclSyntax(
-        bindingSpecifier: .keyword(.let),
-        bindings: PatternBindingListSyntax([
-            PatternBindingSyntax(
-                pattern: IdentifierPatternSyntax(identifier: .identifier("container")),
-                typeAnnotation: TypeAnnotationSyntax(type: containerType)
-            )
-        ])
-    )
-    let bodyDecl = FunctionDeclSyntax(
-        modifiers: accessLevel,
-        name: .identifier("body"),
-        signature: FunctionSignatureSyntax(
-            parameterClause: FunctionParameterClauseSyntax(
-                parameters: FunctionParameterListSyntax([
-                    FunctionParameterSyntax(
-                        firstName: .identifier("content"),
-                        colon: .colonToken(),
-                        type: TypeSyntax(IdentifierTypeSyntax(name: .identifier("Content")))
-                    )
-                ])
-            ),
-            returnClause: ReturnClauseSyntax(type: TypeSyntax(stringLiteral: "some SwiftUI.View"))
-        ),
-        body: CodeBlockSyntax(
-            statements: CodeBlockItemListSyntax([
-                CodeBlockItemSyntax(
-                    item: .stmt(
-                        StmtSyntax(
-                            ExpressionStmtSyntax(
-                                expression: makeEnvironmentBridgeBodyExpr(mappings: mappings)
-                            )
-                        )
-                    )
-                )
-            ])
-        )
-    )
-
-    return StructDeclSyntax(
-        modifiers: accessLevel,
-        name: .identifier("_InnoDIEnvironmentBridgeModifier"),
-        inheritanceClause: InheritanceClauseSyntax(
-            inheritedTypes: InheritedTypeListSyntax([
-                InheritedTypeSyntax(type: TypeSyntax(stringLiteral: "SwiftUI.ViewModifier"))
-            ])
-        ),
-        memberBlock: MemberBlockSyntax(
-            members: MemberBlockItemListSyntax([
-                MemberBlockItemSyntax(decl: DeclSyntax(containerDecl)),
-                MemberBlockItemSyntax(decl: DeclSyntax(bodyDecl)),
-            ])
-        )
-    )
-}
-
-private func makeEnvironmentBridgeHelperDecl(
-    accessLevel: DeclModifierListSyntax
-) -> FunctionDeclSyntax {
-    let modifierType = TypeSyntax(IdentifierTypeSyntax(name: .identifier("_InnoDIEnvironmentBridgeModifier")))
-    let callExpr = ExprSyntax(
-        FunctionCallExprSyntax(
-            calledExpression: ExprSyntax(
-                DeclReferenceExprSyntax(baseName: .identifier("_InnoDIEnvironmentBridgeModifier"))
-            ),
-            leftParen: .leftParenToken(),
-            arguments: LabeledExprListSyntax([
-                LabeledExprSyntax(
-                    label: .identifier("container"),
-                    colon: .colonToken(),
-                    expression: ExprSyntax(DeclReferenceExprSyntax(baseName: .keyword(.self)))
-                )
-            ]),
-            rightParen: .rightParenToken()
-        )
-    )
-
-    return FunctionDeclSyntax(
-        attributes: AttributeListSyntax([
-            .attribute(
-                AttributeSyntax(attributeName: IdentifierTypeSyntax(name: .identifier("MainActor")))
-            )
-        ]),
-        modifiers: accessLevel,
-        name: .identifier("_innodiEnvironmentBridgeModifier"),
-        signature: FunctionSignatureSyntax(
-            parameterClause: FunctionParameterClauseSyntax(parameters: []),
-            returnClause: ReturnClauseSyntax(type: modifierType)
-        ),
-        body: CodeBlockSyntax(
-            statements: CodeBlockItemListSyntax([
-                CodeBlockItemSyntax(
-                    item: .stmt(StmtSyntax(ExpressionStmtSyntax(expression: callExpr)))
-                )
-            ])
-        )
-    )
-}
-
-private func makeEnvironmentBridgeBodyExpr(
-    mappings: [EnvironmentBridgeMappingInfo]
-) -> ExprSyntax {
-    mappings.reduce(
-        ExprSyntax(DeclReferenceExprSyntax(baseName: .identifier("content")))
-    ) { partialResult, mapping in
-        let memberAccess = ExprSyntax(
-            MemberAccessExprSyntax(
-                base: partialResult,
-                declName: DeclReferenceExprSyntax(baseName: .identifier("environment"))
-            )
-        )
-        let containerMember = ExprSyntax(
-            MemberAccessExprSyntax(
-                base: ExprSyntax(DeclReferenceExprSyntax(baseName: .identifier("container"))),
-                declName: DeclReferenceExprSyntax(baseName: .identifier(mapping.memberName))
-            )
-        )
-
-        return ExprSyntax(
-            FunctionCallExprSyntax(
-                calledExpression: memberAccess,
-                leftParen: .leftParenToken(),
-                arguments: LabeledExprListSyntax([
-                    LabeledExprSyntax(
-                        expression: mapping.environmentKeyPath,
-                        trailingComma: .commaToken()
-                    ),
-                    LabeledExprSyntax(expression: containerMember),
-                ]),
-                rightParen: .rightParenToken()
-            )
-        )
-    }
-}
-
-private func parseFeatureRootAttribute(_ attribute: AttributeSyntax) -> FeatureRootAttributeInfo? {
-    guard let arguments = attribute.arguments?.as(LabeledExprListSyntax.self) else {
-        return nil
-    }
-
-    var rootViewTypeName: String?
-    var alias: String?
-
-    for argument in arguments {
-        if let label = argument.label?.text {
-            if label == "as" {
-                alias = stringLiteralValue(argument.expression)
-            }
-            continue
-        }
-
-        if let memberAccess = argument.expression.as(MemberAccessExprSyntax.self),
-           memberAccess.declName.baseName.text == "self",
-           let base = memberAccess.base {
-            rootViewTypeName = base.trimmedDescription
-        } else if argument.expression.trimmedDescription.hasSuffix(".self") {
-            rootViewTypeName = String(argument.expression.trimmedDescription.dropLast(5))
-        }
-    }
-
-    guard let rootViewTypeName else {
-        return nil
-    }
-
-    return FeatureRootAttributeInfo(rootViewTypeName: rootViewTypeName, alias: alias)
-}
-
-private func featureRootHelperName(propertyName: String, alias: String?) -> String {
-    if let alias, !alias.isEmpty {
-        return "\(alias)RootView"
-    }
-    return "\(propertyName)RootView"
-}
-
-private func featureRootHelperConflicts(
-    helperName: String,
-    currentAttribute: AttributeSyntax,
-    in declaration: any DeclGroupSyntax
-) -> Bool {
-    for member in declaration.memberBlock.members {
-        let decl = member.decl
-
-        if let functionDecl = decl.as(FunctionDeclSyntax.self),
-           functionDecl.name.text == helperName {
-            return true
-        }
-
-        if let variableDecl = decl.as(VariableDeclSyntax.self) {
-            if variableDecl.bindings.contains(where: {
-                $0.pattern.as(IdentifierPatternSyntax.self)?.identifier.text == helperName
-            }) {
-                return true
-            }
-
-            for attribute in featureRootAttributes(in: variableDecl.attributes) {
-                guard let info = parseFeatureRootAttribute(attribute) else {
-                    continue
-                }
-                if let alias = info.alias, !isValidFeatureRootAlias(alias) {
-                    continue
-                }
-
-                let propertyName = variableDecl.bindings.first?
-                    .pattern.as(IdentifierPatternSyntax.self)?
-                    .identifier.text
-                    ?? ""
-                let candidateName = featureRootHelperName(propertyName: propertyName, alias: info.alias)
-                if candidateName == helperName,
-                   attribute != currentAttribute,
-                   attributeSortKey(attribute) < attributeSortKey(currentAttribute) {
-                    return true
-                }
-            }
-        }
-    }
-
-    return false
-}
-
-private func keyPathComponentName(from expression: ExprSyntax) -> String? {
-    expression.as(KeyPathExprSyntax.self)?
-        .components
-        .last?
-        .component
-        .as(KeyPathPropertyComponentSyntax.self)?
-        .declName
-        .baseName
-        .text
-}
-
-private func containerMemberNames(in declaration: some DeclGroupSyntax) -> [String] {
+internal func containerMemberNames(in declaration: some DeclGroupSyntax) -> [String] {
     declaration.memberBlock.members.compactMap { member in
         guard let variableDecl = member.decl.as(VariableDeclSyntax.self) else {
             return nil
@@ -546,7 +32,7 @@ private func containerMemberNames(in declaration: some DeclGroupSyntax) -> [Stri
     }
 }
 
-private func nominalTypeSyntax(for declaration: some DeclGroupSyntax) -> TypeSyntax? {
+internal func nominalTypeSyntax(for declaration: some DeclGroupSyntax) -> TypeSyntax? {
     if let structDecl = declaration.as(StructDeclSyntax.self) {
         return nominalTypeSyntax(name: structDecl.name.text, genericParameterClause: structDecl.genericParameterClause)
     }
@@ -562,7 +48,7 @@ private func nominalTypeSyntax(for declaration: some DeclGroupSyntax) -> TypeSyn
     return nil
 }
 
-private func nominalTypeSyntax(
+internal func nominalTypeSyntax(
     name: String,
     genericParameterClause: GenericParameterClauseSyntax?
 ) -> TypeSyntax {
@@ -586,7 +72,9 @@ private func nominalTypeSyntax(
     )
 }
 
-private func accessLevelModifierText(for modifiers: DeclModifierListSyntax?) -> String {
+// MARK: - Access level mapping
+
+internal func accessLevelModifierText(for modifiers: DeclModifierListSyntax?) -> String {
     guard let modifiers else {
         return ""
     }
@@ -611,7 +99,7 @@ private func accessLevelModifierText(for modifiers: DeclModifierListSyntax?) -> 
     return ""
 }
 
-private func accessLevelModifiers(for modifiers: DeclModifierListSyntax?) -> DeclModifierListSyntax {
+internal func accessLevelModifiers(for modifiers: DeclModifierListSyntax?) -> DeclModifierListSyntax {
     let accessLevel = accessLevelModifierText(for: modifiers).trimmingCharacters(in: .whitespaces)
     guard !accessLevel.isEmpty else {
         return DeclModifierListSyntax([])
@@ -638,7 +126,9 @@ private func accessLevelModifiers(for modifiers: DeclModifierListSyntax?) -> Dec
     ])
 }
 
-private func hasAttribute(named name: String, in attributes: AttributeListSyntax?) -> Bool {
+// MARK: - Attribute lookup
+
+internal func hasAttribute(named name: String, in attributes: AttributeListSyntax?) -> Bool {
     findAttribute(
         named: name,
         allowingQualifiedModules: ["InnoDI"],
@@ -646,7 +136,7 @@ private func hasAttribute(named name: String, in attributes: AttributeListSyntax
     ) != nil
 }
 
-private func featureRootAttributes(in attributes: AttributeListSyntax?) -> [AttributeSyntax] {
+internal func featureRootAttributes(in attributes: AttributeListSyntax?) -> [AttributeSyntax] {
     guard let attributes else {
         return []
     }
@@ -663,7 +153,9 @@ private func featureRootAttributes(in attributes: AttributeListSyntax?) -> [Attr
     }
 }
 
-private func isValidFeatureRootAlias(_ alias: String) -> Bool {
+// MARK: - FeatureRoot alias validation
+
+internal func isValidFeatureRootAlias(_ alias: String) -> Bool {
     guard !alias.isEmpty else {
         return false
     }
@@ -682,7 +174,9 @@ private func isSwiftIdentifierBody(_ scalar: UnicodeScalar) -> Bool {
     isSwiftIdentifierHead(scalar) || CharacterSet.decimalDigits.contains(scalar)
 }
 
-private func stringLiteralValue(_ expression: ExprSyntax) -> String? {
+// MARK: - Small AST utilities
+
+internal func stringLiteralValue(_ expression: ExprSyntax) -> String? {
     let text = expression.trimmedDescription
     guard text.count >= 2, text.first == "\"", text.last == "\"" else {
         return nil
@@ -690,11 +184,11 @@ private func stringLiteralValue(_ expression: ExprSyntax) -> String? {
     return String(text.dropFirst().dropLast())
 }
 
-private func attributeSortKey(_ attribute: AttributeSyntax) -> Int {
+internal func attributeSortKey(_ attribute: AttributeSyntax) -> Int {
     attribute.positionAfterSkippingLeadingTrivia.utf8Offset
 }
 
-private func enclosingDeclGroup(containing syntax: Syntax) -> (any DeclGroupSyntax)? {
+internal func enclosingDeclGroup(containing syntax: Syntax) -> (any DeclGroupSyntax)? {
     var current = syntax.parent
     while let node = current {
         if let structDecl = node.as(StructDeclSyntax.self) {
@@ -714,7 +208,7 @@ private func enclosingDeclGroup(containing syntax: Syntax) -> (any DeclGroupSynt
     return nil
 }
 
-private func enclosingDeclModifiers(containing syntax: Syntax) -> DeclModifierListSyntax? {
+internal func enclosingDeclModifiers(containing syntax: Syntax) -> DeclModifierListSyntax? {
     if let enclosingDecl = enclosingDeclGroup(containing: syntax) {
         return enclosingDecl.modifiers
     }
